Snap/Cursor Scrippting
Added a snap script to have the attached object snap to designated positions. Ensured it was all called from cursor script and that visual changes to cursor continued to occur. Beautified the scene with colors to make the cursor easier to view. Ensured that objects snap back to their starting positions if not snapped to desired position and they won't rotate or have additional velocity once let go. Movement post dragging may need to be worked on more as well to be safe and still need to make sure objects aren't stacked to the same position unless it's intended.