Cursor Right Clicking & Refactor
Added some basic right clicking functionality that's similar to the current dragging feature. Will work on it more to display details of the selected cooking container as work on containers is done. Refactored to script to reintroduce the input system to the code and ensure that right clicking works with a layer and interface for details. Involved writing an interface for these details and creating a layer to use.